requirements.txt
python-dotenv
langchain
langchain_community
openai
faiss-cpu
python-docx
